Moved remaining validators in validation.js to own file and created class for Subscription Validation. Updated logic to use this.
